+++ src/CavemanTools/Web/Security/UserBasicRights.cs
@@ -5,14 +5,14 @@ namespace CavemanTools.Web.Security
         /// <summary>
         /// Values is 0
         /// </summary>
-        public const byte None = 0;//this should not be changed
+        public const ushort None = 0;//this should not be changed
         /// <summary>
         /// Value is 1
         /// </summary>
-        public const byte Administration = 1;
+        public const ushort Administration = 1;
         /// <summary>
         /// Value is 2
         /// </summary>
-        public const byte Login = 2;           
+        public const ushort Login = 2;           
     }
 }
